青福换电

+++ src/main/java/com/xiliulou/electricity/controller/admin/JsonAdminElectricityMemberCardOrderController.java
@@ -68,7 +68,7 @@ public class JsonAdminElectricityMemberCardOrderController {
 			return R.fail("ELECTRICITY.0001", "未找到用户");
 		}
 
-		Integer franchiseeId=null;
+		Long franchiseeId=null;
 		if (!Objects.equals(user.getType(), User.TYPE_USER_SUPER)
 				&& !Objects.equals(user.getType(), User.TYPE_USER_OPERATE)) {
 			//加盟商
@@ -117,7 +117,7 @@ public class JsonAdminElectricityMemberCardOrderController {
 			return R.fail("ELECTRICITY.0001", "未找到用户");
 		}
 
-		Integer franchiseeId=null;
+		Long franchiseeId=null;
 		if (!Objects.equals(user.getType(), User.TYPE_USER_SUPER)
 				&& !Objects.equals(user.getType(), User.TYPE_USER_OPERATE)) {
 			//加盟商
@@ -165,7 +165,7 @@ public class JsonAdminElectricityMemberCardOrderController {
 			throw new CustomBusinessException("查不到订单");
 		}
 
-		Integer franchiseeId=null;
+		Long franchiseeId=null;
 		if (!Objects.equals(user.getType(), User.TYPE_USER_SUPER)
 				&& !Objects.equals(user.getType(), User.TYPE_USER_OPERATE)) {
 			//加盟商
